requirements.txt
django
djangorestframework
django-extensions
Werkzeug
argon2-cffi
pyOpenSSL
